Tests are going to be organized. See #46.
