+++ files/rsum/rsum/home/tests.py
@@ -13,6 +13,7 @@ from home.schema.entry import Entry
 
 import home
 import yaml
+from exam.cv import CVTestCase
 from exam.views import ViewsTestCase
 
 home.CV = CV()
@@ -22,32 +23,6 @@ home.Project = Project()
 home.ProjectItem = ProjectItem()
 home.EntryItem = EntryItem()
 home.Entry = Entry()
-
-class CVTestCase(TestCase):
-    def setUp(self):
-        f = open('/srv/rsum/cvs/abridged.yml')
-        self.abridged = yaml.load(f.read())
-        f.close()
-        f = open('/srv/rsum/cvs/complete.yml')
-        self.complete = yaml.load(f.read())
-        f.close
-
-    def test_save_abridged_cv(self):
-        abridged = self.abridged
-        cv = CV()
-        result = cv.save_cv(abridged, 'abridged', template='acecv')
-
-        self.assertEqual(result.name, 'abridged')
-        self.assertEqual(result.template, 'acecv')
-
-    def test_save_complete_cv(self):
-        complete = self.complete
-        cv = CV()
-        result = cv.save_cv(complete, 'complete', template='acecv')
-    
-        self.assertEqual(result.name, 'complete')
-        self.assertEqual(result.template, 'acecv')
-
 
 class SectionTestCase(TestCase):
     def setUp(self):
